# crates/acn-ctl/openapi.json
{
  "components": {
    "schemas": {
      "Answer": {
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The acn-bench control plane (SPEC 070).",
    "title": "acn-ctl",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/shutdown": {
      "post": {
        "operationId": "shutdown",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "OK"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Wrong content type"
          }
        },
        "summary": "Finish the run in progress and stop (CTL-1)."
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "8080"
        }
      }
    }
  ]
}